Import github.com/FiloSottile/Heartbleed directly.
