+++ SAMPLE.aws-config.json
