+++ bleed_serve_moz.go
@@ -1,5 +1,19 @@
 package main
 
+/* Metric modifications
+
+   Metric:
+   .check              total sites checked
+   .cache.hit          cache hit/miss
+   .cache.miss
+   .cache.vulerable    cached sate
+   .cache.safe
+   .cache.error
+   .site.vulnerable    active check state
+   .site.safe
+   .site.error
+*/
+
 import (
 	"encoding/json"
 	"log"
@@ -8,13 +22,16 @@ import (
 
 	"github.com/docopt/docopt-go"
 
-	bleed "github.com/mozilla-services/Heartbleed/bleed"
-	cache "github.com/mozilla-services/Heartbleed/server/cache"
+	bleed "github.com/FiloSottile/Heartbleed/bleed"
+	"github.com/FiloSottile/Heartbleed/server/cache"
+	mz_metrics "github.com/mozilla-services/Heartbleed/metrics"
 )
 
 var PAYLOAD = []byte("filippo.io/Heartbleed")
 
 var withCache bool
+
+var metrics *mz_metrics.Metrics
 
 type result struct {
 	Code  int    `json:"code"`
@@ -35,9 +52,15 @@ func handleRequest(tgt *bleed.Target, w http.ResponseWriter, r *http.Request, sk
 	var errS string
 	var data string
 
+	var rc_state = []string{"vulnerable", "safe", "error"}
+
 	cacheKey := tgt.Service + "://" + tgt.HostIp
 	if skip {
 		cacheKey += "/skip"
+	}
+
+	if metrics != nil {
+		metrics.Increment("check")
 	}
 
 	var cacheOk bool
@@ -48,6 +71,10 @@ func handleRequest(tgt *bleed.Target, w http.ResponseWriter, r *http.Request, sk
 			errS = cReply.Error
 			data = cReply.Data
 			cacheOk = true
+			if metrics != nil {
+				metrics.Increment("cache.hit")
+				metrics.Increment("cache." + rc_state[rc])
+			}
 		}
 	}
 
@@ -95,9 +122,15 @@ func handleRequest(tgt *bleed.Target, w http.ResponseWriter, r *http.Request, sk
 				log.Printf("%v (%v) - ERROR [%v]", tgt.HostIp, tgt.Service, errS)
 			}
 		}
+		if metrics != nil {
+			metrics.Increment("site." + rc_state[rc])
+		}
 	}
 
 	if withCache && !cacheOk {
+		if metrics != nil {
+			metrics.Increment("cache.miss")
+		}
 		cache.Set(cacheKey, rc, data, errS)
 	}
 
@@ -151,7 +184,7 @@ func bleedQueryHandler(w http.ResponseWriter, r *http.Request) {
 var Usage = `Heartbleed test server.
 
 Usage:
-  HBserver --redir-host=<host> [--listen=<addr:port> --expiry=<duration>]
+  HBserver --redir-host=<host> [--listen=<addr:port> --expiry=<duration> --metric.server=<addr:port> --metric.name=<statsd_name> --cache.name=<name>]
   HBserver -h | --help
   HBserver --version
 
@@ -162,17 +195,69 @@ Options:
                       Uses Go's parse syntax
                       e.g. 10m = 10 minutes, 600s = 600 seconds, 1d = 1 day, etc.
   -h --help           Show this screen.
+  --metric.server ADDR:PORT    Address of the statsd server. If no value
+                      present, no statsd reporting is performed
+  --metric.name NAME  Statsd app prefix name to use [default: heartbleed]
+  --cache.name NAME   Name of the AWS DDB cache to use
   --version           Show version.`
 
 func main() {
-	arguments, _ := docopt.Parse(Usage, nil, true, "HBserver 0.3", false)
+	var err error
+	arguments, err := docopt.Parse(Usage, nil, true, "HBserver 0.3", false)
+	if err != nil {
+		log.Printf("ERROR: %s", err.Error())
+		return
+	}
+
+	var statsd_server, statsd_prefix string
 
 	if arguments["--expiry"] != nil {
 		withCache = true
 	}
 
+	if _, ok := arguments["--cache.name"]; !ok {
+		arguments["--cache.name"] = ""
+	}
+
 	if withCache {
+		if arguments["--cache.name"].(string) != "" {
+			cache.CACHE_TAB = arguments["--cache.name"].(string)
+		}
 		cache.Init(arguments["--expiry"].(string))
+	}
+
+	if _, ok := arguments["--listen"]; !ok {
+		arguments["--listen"] = ":8082"
+	}
+
+	// Metric handling block
+	if arguments["--metric.server"] != nil {
+		if prefix, ok := arguments["metric.name"]; !ok {
+			statsd_prefix = "heartbleed"
+		} else {
+			statsd_prefix = prefix.(string)
+		}
+
+		metrics = mz_metrics.New(statsd_prefix, statsd_server)
+		http.HandleFunc("/metrics", func(w http.ResponseWriter,
+			r *http.Request) {
+			w.Header().Set("Content-Type", "application/json")
+			snapshot := metrics.Snapshot()
+			if snapshot == nil {
+				w.Write([]byte("{}"))
+				return
+			}
+			reply, err := json.Marshal(snapshot)
+			if err != nil {
+				log.Printf("handler ERROR: Could not generate report: %s",
+					err.Error())
+				jer := make(map[string]bool)
+				jer["error"] = true
+				reply, err = json.Marshal(jer)
+			}
+			w.Write(reply)
+			return
+		})
 	}
 
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
@@ -188,7 +273,7 @@ func main() {
 	http.HandleFunc("/bleed/query", bleedQueryHandler)
 
 	log.Printf("Starting server on %s\n", arguments["--listen"].(string))
-	err := http.ListenAndServe(arguments["--listen"].(string), nil)
+	err = http.ListenAndServe(arguments["--listen"].(string), nil)
 	if err != nil {
 		log.Fatal("ListenAndServe: ", err)
 	}
