--- deploy/.gitignore
@@ -1,5 +1,0 @@
-hosts.txt
-*.pem
-inventory.ini
-HBserver
-aws-config.json
