--- deploy/ansible.yml
@@ -1,17 +1,0 @@
-- hosts: bleed
-  remote_user: ubuntu
-  sudo: yes
-  tasks:
-  - apt: pkg=daemontools update_cache=yes cache_valid_time=3600
-  - apt: pkg=daemontools-run update_cache=yes cache_valid_time=3600
-  - file: path=/etc/service/bleed state=directory
-  - file: path=/etc/service/bleed/supervise/ state=directory
-  - copy: src=run.sh dest=/etc/service/bleed/run mode=0755
-  - copy: src=HBserver dest=/home/ubuntu/HBserver
-  - copy: src=aws-config.json dest=/etc/aws-config.json
-  - service: name=svscan state=started
-  - command: pkill -f heartbleed
-    ignore_errors: yes
-  - command: pkill -f HBserver
-    ignore_errors: yes
-  - command: svc -u /etc/service/bleed
