--- deploy/decommission.yml
@@ -1,32 +1,0 @@
-- hosts: bleed-old
-  remote_user: ubuntu
-  vars:
-    logs_host: ec2-54-237-188-98.compute-1.amazonaws.com
-  tasks:
-    - copy: src=bleed_logs.pem dest=.ssh/id_rsa mode=0600
-    - command: svc -d /etc/service/bleed
-      sudo: yes
-    - command: pkill -f heartbleed
-      ignore_errors: yes
-      sudo: yes
-    - command: pkill -f HBserver
-      ignore_errors: yes
-      sudo: yes
-    - shell: ssh-keyscan -t ecdsa,rsa,dsa {{logs_host}} >> ~/.ssh/known_hosts;
-    - command: rsync --partial heartbleed.log ubuntu@{{logs_host}}:bleed/{{inventory_hostname}}.log
-    - command: shutdown -h now
-      sudo: yes
-
-- hosts: ref-old
-  remote_user: ubuntu
-  vars:
-    logs_host: ec2-54-237-188-98.compute-1.amazonaws.com
-  tasks:
-    - copy: src=bleed_logs.pem dest=.ssh/id_rsa mode=0600
-    - command: pkill -f logger
-      ignore_errors: yes
-      sudo: yes
-    - shell: ssh-keyscan -t ecdsa,rsa,dsa {{logs_host}} >> ~/.ssh/known_hosts;
-    - command: rsync logger_linux_amd64/referrals.log ubuntu@{{logs_host}}:ref/{{inventory_hostname}}.log
-    - command: shutdown -h now
-      sudo: yes
